Refactor Mist core functions
Refactor the core functions to make two changes. First, DRY it up by
moving the code that does the enqueueing and checking for currentUser
into one function. Second, put the check for the currentUser inside the
operation rather than outside it, so that we don’t check for the User
until we’re about to actually do the work.