Refactor Mist core functions; update dependent code
Refactor the core functions to make two changes. First, DRY it up by
moving the code that does the enqueueing and checking for currentUser
into one function. Second, put the check for the currentUser inside the
operation rather than outside it, so that we don’t check for the User
until we’re about to actually do the work.

In addition, create a Mist singleton that we use internally so that our
instance of LocalDataCoordinator doesn’t disappear out from under us
(not sure this is necessary).

Finally, create a new function to give the sync code async access to
the cached changes so that we actually have them when we go to do a
push.